+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add more information to parsed tuc types
 
 ## 1.3.0 - 2020-11-03
 - Use `Tuc.DomainResolver` as a external library